# src/main/resources/openapi/domestic-payment-api.yaml
openapi: 3.0.3
info:
  title: Domestic RTP Payment Service API
  version: 0.1.0
  description: External-facing domestic real-time payment initiation and status query API.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
paths:
  /v1/domestic-payments:
    post:
      operationId: CreateDomesticPayment
      summary: Create a domestic real-time payment
      description: Accepts one domestic real-time payment instruction for asynchronous processing.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/MockScenario'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDomesticPaymentRequest'
            examples:
              success:
                value:
                  debtorAccount:
                    bankCode: CIBMYKL
                    accountNumber: '1234567890'
                    accountName: ACME Operating Account
                  creditorAccount:
                    bankCode: CIBMYKL
                    accountNumber: '9876543210'
                    accountName: Supplier Sdn Bhd
                  amount:
                    currency: MYR
                    value: '1250.00'
                  paymentReference: INV-2026-0001
                  remittanceInformation: Invoice payment
                  requestedExecutionDate: '2026-05-09'
      responses:
        '202':
          description: Payment request accepted for asynchronous processing
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/domestic-payments/{paymentId}:
    get:
      operationId: GetDomesticPaymentStatus
      summary: Get domestic payment status
      description: Returns the current known status for a payment visible to the authenticated client.
      security:
        - bearerAuth: []
      parameters:
        - name: paymentId
          in: path
          required: true
          description: Payment ID returned by payment creation.
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Current payment status
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentStatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-supplied key scoped to the authenticated client for idempotent payment creation.
      schema:
        type: string
        minLength: 1
        maxLength: 128
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: End-to-end tracing identifier. Generated by the API when absent.
      schema:
        type: string
        minLength: 1
        maxLength: 128
    MockScenario:
      name: X-Mock-Scenario
      in: header
      required: false
      description: Local/test-only downstream mock scenario selector. Not part of production payment semantics.
      schema:
        type: string
        enum:
          - success
          - rejection
          - timeout
          - internal_failure
  headers:
    CorrelationId:
      description: Correlation ID used for request tracing.
      schema:
        type: string
  responses:
    BadRequest:
      description: Malformed request, validation failure, or missing required header
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, expired, invalid, or unverifiable JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated client lacks the required scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Payment not found or not visible to the authenticated client
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    IdempotencyConflict:
      description: Idempotency key was reused with a different request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Semantically invalid payment instruction
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Unexpected API failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    CreateDomesticPaymentRequest:
      type: object
      additionalProperties: false
      required:
        - debtorAccount
        - creditorAccount
        - amount
        - paymentReference
      properties:
        debtorAccount:
          $ref: '#/components/schemas/AccountReference'
        creditorAccount:
          $ref: '#/components/schemas/AccountReference'
        amount:
          $ref: '#/components/schemas/Money'
        paymentReference:
          type: string
          minLength: 1
          maxLength: 140
        remittanceInformation:
          type: string
          maxLength: 280
        requestedExecutionDate:
          type: string
          format: date
    AccountReference:
      type: object
      additionalProperties: false
      required:
        - bankCode
        - accountNumber
        - accountName
      properties:
        bankCode:
          type: string
          minLength: 1
          maxLength: 35
        accountNumber:
          type: string
          minLength: 1
          maxLength: 34
        accountName:
          type: string
          minLength: 1
          maxLength: 140
    Money:
      type: object
      additionalProperties: false
      required:
        - currency
        - value
      properties:
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: MYR
        value:
          type: string
          pattern: '^[0-9]+(\.[0-9]{1,2})?$'
          example: '1250.00'
    PaymentResponse:
      type: object
      required:
        - paymentId
        - status
        - createdAt
        - updatedAt
        - correlationId
        - links
      properties:
        paymentId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/PaymentStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        correlationId:
          type: string
        links:
          type: object
          required:
            - status
          properties:
            status:
              type: string
    PaymentStatusResponse:
      type: object
      required:
        - paymentId
        - status
        - createdAt
        - updatedAt
        - correlationId
        - links
      properties:
        paymentId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/PaymentStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        correlationId:
          type: string
        reason:
          $ref: '#/components/schemas/PaymentReason'
        links:
          type: object
          required:
            - self
          properties:
            self:
              type: string
    PaymentStatus:
      type: string
      enum:
        - ACCEPTED
        - PROCESSING
        - COMPLETED
        - REJECTED
        - FAILED
        - TIMEOUT
    PaymentReason:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - status
        - correlationId
      properties:
        code:
          type: string
        message:
          type: string
        status:
          type: integer
          format: int32
        correlationId:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorDetail'
    ValidationErrorDetail:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string